KB2919442, KB2919355: recognize WU_E_NOT_APPLICABLE exit code
If a hotfix is superseded and has never been installed on the machine,
Get-HotFix will not return it. Upon an attempt to install the hotfix,
wusa.exe will return 0x80240017 = WU_E_NOT_APPLICABLE [1].
This needs to be treated as a success.

Fixes installation of KB2919442 on a freshly created Azure 2012 R2 VM.

[1] https://technet.microsoft.com/en-us/library/cc720442(v=ws.10).aspx